requirements-dev.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
